Move LetterBox to GuessBars.py to avoid circular import
